requirements.txt
prefect
openai-whisper
ffmpeg-python
